+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- New permission `approve_grant_request`
+- [TD-4076] store roles with permissions in cache
+
 ## [4.26.1] 2021-08-18
 
 ### Fixed